Speed up check by analysing each line only once
This reduces the amount of work done per line significantly,
primarily by not searching through it in duplicate for both checks.

While the speedup is hard to notice on small codebases, this offers
a speedup of about 2× on a large codebase. (Tested on a 80k line
codebase, where the time for this checker drops from ~21s to ~11s).